# Add tracing, stderr logging, and parallel image alterations