--- a/libwayshot/src/dispatch.rs
+++ b/libwayshot/src/dispatch.rs
@@ -68,7 +68,7 @@ impl Dispatch<WlRegistry, ()> for OutputCaptureState {
                         },
                     });
                 } else {
-                    log::error!("Ignoring a wl_output with version < 4.");
+                    tracing::error!("Ignoring a wl_output with version < 4.");
                 }
             }
         }
@@ -128,12 +128,12 @@ impl Dispatch<ZxdgOutputV1, usize> for OutputCaptureState {
             zxdg_output_v1::Event::LogicalPosition { x, y } => {
                 output_info.dimensions.x = x;
                 output_info.dimensions.y = y;
-                log::debug!("Logical position event fired!");
+                tracing::debug!("Logical position event fired!");
             }
             zxdg_output_v1::Event::LogicalSize { width, height } => {
                 output_info.dimensions.width = width;
                 output_info.dimensions.height = height;
-                log::debug!("Logical size event fired!");
+                tracing::debug!("Logical size event fired!");
             }
             _ => {}
         };
@@ -171,7 +171,7 @@ impl Dispatch<ZwlrScreencopyFrameV1, ()> for CaptureFrameState {
                 height,
                 stride,
             } => {
-                log::debug!("Received Buffer event");
+                tracing::debug!("Received Buffer event");
                 if let Value(f) = format {
                     frame.formats.push(FrameFormat {
                         format: f,
@@ -180,31 +180,31 @@ impl Dispatch<ZwlrScreencopyFrameV1, ()> for CaptureFrameState {
                         stride,
                     })
                 } else {
-                    log::debug!("Received Buffer event with unidentified format");
+                    tracing::debug!("Received Buffer event with unidentified format");
                     exit(1);
                 }
             }
             zwlr_screencopy_frame_v1::Event::Flags { .. } => {
-                log::debug!("Received Flags event");
+                tracing::debug!("Received Flags event");
             }
             zwlr_screencopy_frame_v1::Event::Ready { .. } => {
                 // If the frame is successfully copied, a “flags” and a “ready” events are sent. Otherwise, a “failed” event is sent.
                 // This is useful when we call .copy on the frame object.
-                log::debug!("Received Ready event");
+                tracing::debug!("Received Ready event");
                 frame.state.replace(FrameState::Finished);
             }
             zwlr_screencopy_frame_v1::Event::Failed => {
-                log::debug!("Received Failed event");
+                tracing::debug!("Received Failed event");
                 frame.state.replace(FrameState::Failed);
             }
             zwlr_screencopy_frame_v1::Event::Damage { .. } => {
-                log::debug!("Received Damage event");
+                tracing::debug!("Received Damage event");
             }
             zwlr_screencopy_frame_v1::Event::LinuxDmabuf { .. } => {
-                log::debug!("Received LinuxDmaBuf event");
+                tracing::debug!("Received LinuxDmaBuf event");
             }
             zwlr_screencopy_frame_v1::Event::BufferDone => {
-                log::debug!("Received bufferdone event");
+                tracing::debug!("Received bufferdone event");
                 frame.buffer_done.store(true, Ordering::SeqCst);
             }
             _ => unreachable!(),