--- a/libwayshot/src/lib.rs
+++ b/libwayshot/src/lib.rs
@@ -140,13 +140,15 @@ impl WayshotConnection {
         event_queue.roundtrip(&mut state)?;
         event_queue.roundtrip(&mut state)?;
 
-        let mut xdg_outputs: Vec<ZxdgOutputV1> = Vec::new();
-
         // We loop over each output and request its position data.
-        for (index, output) in state.outputs.clone().iter().enumerate() {
-            let xdg_output = zxdg_output_manager.get_xdg_output(&output.wl_output, &qh, index);
-            xdg_outputs.push(xdg_output);
-        }
+        let xdg_outputs: Vec<ZxdgOutputV1> = state
+            .outputs
+            .iter()
+            .enumerate()
+            .map(|(index, output)| {
+                zxdg_output_manager.get_xdg_output(&output.wl_output, &qh, index)
+            })
+            .collect();
 
         event_queue.roundtrip(&mut state)?;
 
@@ -376,24 +378,28 @@ impl WayshotConnection {
         capture_region: CaptureRegion,
         cursor_overlay: i32,
     ) -> Result<Frame> {
-        let outputs = self.get_all_outputs();
-        let mut intersecting_outputs: Vec<IntersectingOutput> = Vec::new();
-        for output in outputs.iter() {
-            let x1: i32 = cmp::max(output.dimensions.x, capture_region.x_coordinate);
-            let y1: i32 = cmp::max(output.dimensions.y, capture_region.y_coordinate);
-            let x2: i32 = cmp::min(
-                output.dimensions.x + output.dimensions.width,
-                capture_region.x_coordinate + capture_region.width,
-            );
-            let y2: i32 = cmp::min(
-                output.dimensions.y + output.dimensions.height,
-                capture_region.y_coordinate + capture_region.height,
-            );
+        let frame_copies = self
+            .get_all_outputs()
+            .into_par_iter()
+            .filter_map(|output| {
+                let x1: i32 = cmp::max(output.dimensions.x, capture_region.x_coordinate);
+                let y1: i32 = cmp::max(output.dimensions.y, capture_region.y_coordinate);
+                let x2: i32 = cmp::min(
+                    output.dimensions.x + output.dimensions.width,
+                    capture_region.x_coordinate + capture_region.width,
+                );
+                let y2: i32 = cmp::min(
+                    output.dimensions.y + output.dimensions.height,
+                    capture_region.y_coordinate + capture_region.height,
+                );
 
-            let width = x2 - x1;
-            let height = y2 - y1;
+                let width = x2 - x1;
+                let height = y2 - y1;
 
-            if !(width <= 0 || height <= 0) {
+                if width <= 0 || height <= 0 {
+                    return None;
+                }
+
                 let true_x = capture_region.x_coordinate - output.dimensions.x;
                 let true_y = capture_region.y_coordinate - output.dimensions.y;
                 let true_region = CaptureRegion {
@@ -402,20 +408,12 @@ impl WayshotConnection {
                     width: capture_region.width,
                     height: capture_region.height,
                 };
-                intersecting_outputs.push(IntersectingOutput {
+                Some(IntersectingOutput {
                     output: output.wl_output.clone(),
                     region: true_region,
                     transform: output.transform,
-                });
-            }
-        }
-        if intersecting_outputs.is_empty() {
-            tracing::error!("Provided capture region doesn't intersect with any outputs!");
-            exit(1);
-        }
-
-        let frame_copies = intersecting_outputs
-            .par_iter()
+                })
+            })
             .map(|intersecting_output| {
                 self.capture_output_frame(
                     cursor_overlay,
@@ -442,6 +440,7 @@ impl WayshotConnection {
             .par_iter()
             .map(|frame_copy| -> Result<_> {
                 let image = frame_copy.try_into()?;
+                tracing::info!("ROTATING");
                 Ok(image_util::rotate_image_buffer(
                     image,
                     frame_copy.transform,
@@ -450,12 +449,14 @@ impl WayshotConnection {
                 ))
             })
             .try_reduce_with(|mut merged_image, image| {
+                tracing::info!("OVERLAYING");
                 overlay(&mut merged_image, &image, 0, 0);
                 Ok(merged_image)
             })
-            // If there are no frame copies, this will return None, but that
-            // shouldn't happen. Return error of no outputs just in case.
-            .ok_or(Error::NoOutputs)?
+            .ok_or_else(|| {
+                tracing::error!("Provided capture region doesn't intersect with any outputs!");
+                Error::NoOutputs
+            })?
     }
 
     /// shot one ouput