… Cargo.lock
… libwayshot/Cargo.toml
… wayshot/Cargo.toml