@Shinyzenith 2023-10-12
Hi! thank you for the patch. I haven't gone through all of it just yet. Just to be sure, rayon was replaced with scoped threads right?

I spotted some rayon artifacts in `use` and `Cargo.lock` during a quick overlook. Is that intentional?