@Shinyzenith 2023-10-12
Apart from those two tiny nit-picks, everything looks fine and the PR performs really well! Thank you for the PR!